kbuild: add script and target to generate apk package
apk is the package manager by Alpine Linux (not to be confused with apk
from Android systems).

The goal is to improve the workflow for developers of postmarketOS,
a derivative of Alpine Linux focused on mobile and embedded devices,
to contribute to the Linux kernel. By installing the kernel as a apk, the
required triggers are being executed, to re-configure the initfs and the
bootloader.

The APKBUILD file is similar to the PKGBUILD file by pacman, as their
build system is built similar.

Signed-off-by: Achill Gilgenast <fossdd@pwned.life>